% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Standard Frequency Tables in R}
\usage{
frequencies(
  x = vector(),
  n.classes = nclass.Sturges(x),
  sort.decreasing = NULL,
  use.hist.classes = FALSE,
  count.na = TRUE,
  n.digits = 2,
  show.totals = TRUE,
  show.percent = FALSE,
  as.markdown = FALSE,
  show.relative = TRUE,
  show.cumulative = TRUE,
  show.rel.cumulative = TRUE,
  show.frequencies = TRUE,
  compare.valids = FALSE
)
}
\arguments{
\item{x}{a numeric or categorical variable}

\item{n.classes}{the number of classes for the numeric variable}

\item{sort.decreasing}{sort frequencies decreasingly}

\item{use.hist.classes}{for numeric variables, use the default classes for histograms}

\item{count.na}{include NA values as part of the classes to be counted in}

\item{n.digits}{number of decimal places to display for relative frequencies}

\item{show.totals}{show the sum of frequencies}

\item{show.percent}{show relative frequencies as percentages}

\item{as.markdown}{return the frequency in RMarkdown format}

\item{show.relative}{show the column for relative frequencies}

\item{show.cumulative}{show the column for cumulative frequencies}

\item{show.rel.cumulative}{show the column for relative cumulative frequencies}

\item{show.frequencies}{show counts per class}

\item{compare.valids}{show an additional column for each column type to compare valid vs. NA values. It overrides na.count}
}
\value{
a data.frame with one character column and the rest as numeric values
}
\description{
Standard Frequency Tables in R
}
\examples{
frequencies(x = airquality$Ozone)
}
