% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Standard Frequency Tables in R}
\usage{
frequencies(
  x = vector(),
  n.classes = nclass.Sturges(x),
  sort.decreasing = NULL,
  use.hist.classes = FALSE,
  count.na = TRUE,
  n.digits = 2,
  show.totals = TRUE,
  show.percent = FALSE,
  as.markdown = FALSE,
  show.relative = TRUE,
  show.cumulative = TRUE,
  show.rel.cumulative = TRUE,
  show.frequencies = TRUE,
  compare.valids = FALSE,
  x.as.categorical = FALSE,
  svy.design = NULL
)
}
\arguments{
\item{x}{vector. a numeric or categorical variable}

\item{n.classes}{numeric. the number of classes for the numeric variable}

\item{sort.decreasing}{NULL or logical. sort frequencies decreasingly, increasingly according to frequency counts. The default, NULL, sorts frequencies according with the classes.}

\item{use.hist.classes}{logical. uses the default classes for histograms, allowing the frequency table of numeric values to match the default graphical bins for an histogram.}

\item{count.na}{logical. include NA values as part of the classes to be counted. If the vector has no missing values it will report 0 occurrences when TRUE is selected-}

\item{n.digits}{integer. number of decimal places to display for percentage frequencies. The number of decimals shown for relative frequencies will be digits + 2.}

\item{show.totals}{logical. Show the sum of frequencies and relative/percentage frequencies.}

\item{show.percent}{logical. show relative frequencies as percentage frequencies.}

\item{as.markdown}{logical. Return the frequency table in RMarkdown format, uses the pander:: package.}

\item{show.relative}{logical. Show or hide the column for relative frequencies, default is TRUE.}

\item{show.cumulative}{logical. Show or hide the column for cumulative frequencies, default is TRUE.}

\item{show.rel.cumulative}{logical. Show or hide the column for relative cumulative frequencies, default is TRUE.}

\item{show.frequencies}{logical. Show or hide the counts per class, default is TRUE.}

\item{compare.valids}{logical. Show or hide an additional column for each column type to compare valid vs. NA values, default is FALSE. It overrides selected option for na.count argument.}

\item{x.as.categorical}{logical. Force the presentation of a continuous variable as a categorical one.}

\item{svy.design}{survey.design object. Survey design from survey:: package}
}
\value{
a data.frame with one character column and the rest as numeric values
}
\description{
Standard Frequency Tables in R
}
\examples{
frequencies(x = airquality$Ozone)
}
